Added several comments throughout for clarity. Changed the names of many variables and methods for clarity.
Created the Creatures package which contains the CreatureNode class and
individual creature classes. The CreatureNode abstract class is extended
by all Creature classes and acts as a template which provides general
creature functionality such as health points and refresh frame rate, and
also allows the creature objects to be linked in the fashion of a linked
list. The Monster abstract class and Gleeb classes were replaced with
this update.

I chose to structure the creature classes in a way that allows them to
be linked simply because I wanted to be able to "scroll" through the
creatures from the interface. I could have accomplished this easier by
collecting the creature objects in an array and cycling through by their
indexes, but the idea of creating linked nodes that doubled as the
creature objects seemed like fun.