Replaced the Gleeb and Monster classes with the CreatureNode and CreatureList structure. Implemented the Auto, Previous, Next, Repeat buttons.
The creatures are now CreatureNodes and are stored in a linked structure called CreatureList. When enabled, the auto button initiates the "battle" mode as soon as a new creature is deployed. The previous button deploys the creature previous to the current creature in the CreatureList, and the Next button deploys the next (if the current creature has been defeated). When enabled the repeat button will redeploy the current creature after it is defeated.